feat: implement studio pricing calculations and snapshot handling in booking process

+++ src/models/stream_project_booking.js
@@ -56,7 +56,7 @@ module.exports = function(sequelize, DataTypes) {
       allowNull: true
     },
     duration_hours: {
-      type: DataTypes.INTEGER,
+      type: DataTypes.DECIMAL(5, 2),
       allowNull: true
     },
     start_time: {
